requirements.txt
<<<<<<< HEAD
Flask==2.3.3
pymysql==1.1.0
mysql-connector-python==8.3.0
werkzeug==2.3.7
python-dotenv==1.0.1
=======
Flask==3.0.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
click==8.1.7
colorama==0.4.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
protobuf==4.21.12
python-dotenv==1.0.0
blinker==1.7.0 
>>>>>>> 0ac6a7a595adb0b486dc117035bfbae8994ed256
